feat: added deletion, update, addition for chapter entity
